requirements.txt
joblib==1.5.1
opencv_python==4.12.0.88
pandas==2.3.1
Pillow==11.3.0
scikit_learn==1.7.1
timm==1.0.19
torch==2.7.1
torchvision==0.22.1
ultralytics==8.3.174